Allow Using Filename for Torrents Too
We use `/infoHash/fileIdx` in Stremio, users have been complaining for years that we should use the `/infoHash/fileName` so external players can find subs and generally to be able to see the filename.

With this change we add support for `/infoHash/fileName` also, along with the addition of `/infoHash/fileIdx?external=1` which will do a `307` (temporary redirect) to `/infoHash/fileName`.

So from now on, we can just add the `external` query parameter when using external video players or "download" buttons to preserve the filename for users.